python-downloader/requirements.txt
# Cloud Downloader Service Requirements
# Python 3.8+

# Core dependencies
requests>=2.31.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Optional: For enhanced downloading
gdown>=4.7.1  # Google Drive downloads (alternative method)

# Optional: For text extraction
PyMuPDF>=1.23.0  # PDF text extraction (fitz)
pytesseract>=0.3.10  # OCR for images
Pillow>=10.0.0  # Image processing
python-docx>=1.0.0  # Word document extraction

# Optional: For resume parsing
# resume-parser>=0.9.0  # Uncomment if you want built-in resume parsing

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0  # For testing FastAPI
black>=23.0.0  # Code formatting
flake8>=6.0.0  # Linting

# Production dependencies
gunicorn>=21.0.0  # Production WSGI server
python-multipart>=0.0.6  # For file uploads
